+++ test/E2ETest/Microsoft.Test.E2E.AspNet.OData/Build.AspNetCore/Microsoft.Test.E2E.AspNetCore.OData.csproj
@@ -1399,6 +1399,21 @@
     <Compile Include="..\OpenType\TypedTest.cs">
       <Link>OpenType\TypedTest.cs</Link>
     </Compile>
+	<Compile Include="..\BulkOperation\BulkInsertDataModel.cs">
+		<Link>BulkOperation\BulkInsertDataModel.cs</Link>
+	</Compile>
+	<Compile Include="..\BulkOperation\BulkInsertEdmModel.cs">
+		<Link>BulkOperation\BulkInsertEdmModel.cs</Link>
+	</Compile>
+	<Compile Include="..\BulkOperation\BulkInsertTest.cs">
+		<Link>BulkOperation\BulkInsertTest.cs</Link>
+	</Compile>
+	<Compile Include="..\BulkOperation\BulkInsertController.cs">
+		<Link>BulkOperation\BulkInsertController.cs</Link>
+	</Compile>
+	<Compile Include="..\BulkOperation\BulkOperationPatchHandlers.cs">
+		<Link>BulkOperation\BulkOperationPatchHandlers.cs</Link>
+	</Compile>
     <Compile Include="..\ParameterAlias\ParameterAliasDataSource.cs">
       <Link>ParameterAlias\ParameterAliasDataSource.cs</Link>
     </Compile>
