+++ test/E2ETest/Microsoft.Test.E2E.AspNet.OData/BulkOperation/BulkInsertDataModel.cs
@@ -1,16 +1,13 @@
 ﻿// Copyright (c) Microsoft Corporation.  All rights reserved.
 // Licensed under the MIT License.  See License.txt in the project root for license information.
 
-using Microsoft.AspNet.OData.Builder;
-using Purchasing;
 using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.Data.Entity;
+using Microsoft.AspNet.OData.Builder;
 
-namespace Microsoft.Test.E2E.AspNet.OData.BulkInsert1
+namespace Microsoft.Test.E2E.AspNet.OData.BulkInsert
 {
     public class Employee
     {
