+++ test/UnitTest/Microsoft.AspNet.OData.Test/packages.config
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Castle.Core" version="4.2.0" targetFramework="net45" />
-  <package id="EntityFramework" version="6.4.4" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi" version="5.2.2" targetFramework="net45" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.2" targetFramework="net45" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.2" targetFramework="net45" />
