+++ test/UnitTest/Microsoft.AspNetCore.OData.Test/Microsoft.AspNetCore.OData.Test.csproj
@@ -59,15 +59,5 @@
     <Folder Include="Builder\Conventions\" />
     <Folder Include="Properties\" />
   </ItemGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp3.1'">
-    <PackageReference Include="Microsoft.EntityFrameworkCore">
-      <Version>5.0.4</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory">
-      <Version>5.0.4</Version>
-    </PackageReference>
-  </ItemGroup>
-
-	
+  	
 </Project>
